+++ backend/database/storage/bolt/users.go
@@ -75,6 +75,10 @@ func (st usersBackend) Update(user *users.User, actorIsAdmin bool, fields ...str
 
 	if !slices.Contains(fields, "Password") {
 		user.Password = existingUser.Password
+	} else {
+		if existingUser.LockPassword {
+			return fmt.Errorf("password cannot be changed when lock password is enabled")
+		}
 	}
 
 	if !actorIsAdmin {